+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+## [2026.06.20.1452] - 2026-06-20
+
+- Automated build for commit 1792b2a2bcef4f7a7988a2a6fe6c621077ccf618.
+
+## [2026.06.20.1437] - 2026-06-20
+
+- Automated build for commit 1792b2a2bcef4f7a7988a2a6fe6c621077ccf618.
+
+## [2026.06.19.2035] - 2026-06-19
+
+- Automated build for commit 1792b2a2bcef4f7a7988a2a6fe6c621077ccf618.
+
+## [2026.06.19.2022] - 2026-06-19
+
+- Automated build for commit 1792b2a2bcef4f7a7988a2a6fe6c621077ccf618.
+
+## [2026.06.19.2020] - 2026-06-19
+
+- Automated build for commit 1792b2a2bcef4f7a7988a2a6fe6c621077ccf618.
+
+## [2026.06.19.2018] - 2026-06-19
+
+- Automated build for commit 1792b2a2bcef4f7a7988a2a6fe6c621077ccf618.
+
+## [2026.06.19.1643] - 2026-06-19
+
+- Automated build for commit 1792b2a2bcef4f7a7988a2a6fe6c621077ccf618.
+
+## [2026.06.19.1630] - 2026-06-19
+
+- Automated build for commit 1792b2a2bcef4f7a7988a2a6fe6c621077ccf618.
+
 ## [2026.06.19.1357] - 2026-06-19
 
 - Automated build for commit e1160f6f61660cd329724dc0a41cdc6c83cf03f1.